requirements.txt
DonationAlertsHandler==1.1
aiogram==3.13.1
PyYAML==6.0.2
Jinja2==3.1.4
